fix(codex): revalidate owner-local catalog reads
