+++ src/gateway/server-methods/session-catalog-authorization.ts
@@ -1,16 +1,53 @@
+import { normalizeOptionalString } from "@openclaw/normalization-core/string-coerce";
 import {
   ErrorCodes,
   errorShape,
   type SessionCatalogLocator,
 } from "../../../packages/gateway-protocol/src/index.js";
 import type { SessionCatalogProvider } from "../../plugins/session-catalog.js";
+import { resolveAgentIdOrRespondError } from "./agent-id-shared.js";
 import {
   allowProcessHomeFallback,
   createSessionCatalogRequestNodeSnapshot,
   listSessionCatalogProvider,
 } from "./session-catalog-provider-access.js";
-import { isSessionCatalogThreadVisible } from "./session-catalog-visibility.js";
+import {
+  isSessionCatalogThreadVisible,
+  resolveSessionCatalogVisibility,
+} from "./session-catalog-visibility.js";
 import type { GatewayClient, GatewayRequestContext, RespondFn } from "./types.js";
+
+export async function authorizeSessionCatalogRequest(params: {
+  access: "read" | "mutate";
+  request: SessionCatalogLocator & { agentId?: string };
+  provider: SessionCatalogProvider;
+  respond: RespondFn;
+  context: GatewayRequestContext;
+  client: GatewayClient | null;
+}): Promise<
+  | ({ agentId: string } & NonNullable<Awaited<ReturnType<typeof authorizeSessionCatalogThread>>>)
+  | null
+> {
+  const resolvedAgent = resolveAgentIdOrRespondError({
+    rawAgentId: params.request.agentId,
+    respond: params.respond,
+    cfg: params.context.getRuntimeConfig(),
+    normalize: normalizeOptionalString,
+  });
+  if (!resolvedAgent) {
+    return null;
+  }
+  const authorization = await authorizeSessionCatalogThread({
+    access: params.access,
+    agentId: resolvedAgent.agentId,
+    client: params.client,
+    context: params.context,
+    provider: params.provider,
+    request: params.request,
+    respond: params.respond,
+  });
+  return authorization ? { agentId: resolvedAgent.agentId, ...authorization } : null;
+}
 
 export async function authorizeSessionCatalogThread(params: {
   access: "read" | "mutate";
@@ -20,7 +57,10 @@ export async function authorizeSessionCatalogThread(params: {
   provider: SessionCatalogProvider;
   request: SessionCatalogLocator;
   respond: RespondFn;
-}): Promise<{ allowProcessHomeFallback: boolean } | null> {
+}): Promise<{
+  allowProcessHomeFallback: boolean;
+  revalidateBeforePublish?: () => Promise<boolean>;
+} | null> {
   const allowHomeFallback = allowProcessHomeFallback(params.context.logGateway);
   const visible = await isSessionCatalogThreadVisible({
     access: params.access,
@@ -37,7 +77,25 @@ export async function authorizeSessionCatalogThread(params: {
     threadId: params.request.threadId,
   });
   if (visible) {
-    return { allowProcessHomeFallback: allowHomeFallback };
+    const visibility = resolveSessionCatalogVisibility(
+      params.client,
+      params.context.getRuntimeConfig(),
+    );
+    const requiresReadRevalidation =
+      params.access === "read" &&
+      visibility.kind === "restricted-unprofiled" &&
+      visibility.gatewayOwner &&
+      typeof params.provider.audience === "object" &&
+      params.provider.audience.kind === "gateway-owner-local";
+    return {
+      allowProcessHomeFallback: allowHomeFallback,
+      ...(requiresReadRevalidation
+        ? {
+            revalidateBeforePublish: async () =>
+              Boolean(await authorizeSessionCatalogThread(params)),
+          }
+        : {}),
+    };
   }
   params.respond(
     false,
