+++ src/gateway/server/ws-connection/connect-device-pairing.test.ts
@@ -227,6 +227,7 @@ describe("gateway connect pairing exemptions", () => {
       expect(visibility.kind).toBe("unrestricted");
       expect(JSON.parse(visibility.cacheKey)).toEqual({
         admin: true,
+        gatewayOwner: false,
         multipleIdentities: false,
         profileId: null,
         profileAliases: [],
